redid the UserContext, and I merged permission service into Location service to have the Location Context importing from one file
