+++ client/app/contexts/UserContext.tsx
@@ -1,7 +1,6 @@
-import React, { createContext, useContext, useEffect, useState } from "react";
-
+import React, { createContext, useContext, useState } from "react";
 import { UserType } from "../types/User";
-import { generateName } from "@app/utils/scripts";
+import { initializeUser } from "@app/services/UserService";
 
 const UserContext = createContext<UserType | null>(null);
 
@@ -10,17 +9,7 @@ export const useUser = () => {
 };
 
 export const UserProvider = ({ children }: { children: React.ReactNode }) => {
-  const [user, setUser] = useState<UserType>({
-    displayName: "DefaultDisplayName",
-    userIcon: {
-      imagePath: "DefaultImagePath",
-      colorHex: "#fff",
-    },
-  });
-
-  useEffect(() => {
-    user.displayName = generateName()
-  }, [])
+  const [user, setUser] = useState<UserType>(initializeUser); 
 
   return <UserContext.Provider value={user}>{children}</UserContext.Provider>;
 };
